feat: takes Repeat component into it's own seperate file
